fix: abstraction for modal checkout and rename button block
