+++ src/blocks/checkout-button/edit.scss
@@ -1,6 +1,6 @@
 @use '~@wordpress/base-styles/colors' as wp-colors;
 
-.newspack-purchase-button {
+.newspack-checkout-button {
 	&__product-field {
 		&__selected {
 			.components-button {
