+++ src/modal-checkout/block.js
@@ -20,6 +20,8 @@ function domReady( callback ) {
 	document.addEventListener( 'DOMContentLoaded', callback );
 }
 
+const triggers = '.wpbnbd.wpbnbd--platform-wc,.wp-block-newspack-blocks-checkout-button';
+
 let iframeResizeObserver;
 
 function closeCheckout( element ) {
@@ -47,11 +49,9 @@ domReady( () => {
 	const iframe = document.createElement( 'iframe' );
 	iframe.name = iframeName;
 	modalContent.appendChild( iframe );
-	const blocks = document.querySelectorAll(
-		'.wpbnbd.wpbnbd--platform-wc,.wp-block-newspack-blocks-purchase-button'
-	);
-	blocks.forEach( block => {
-		const forms = block.querySelectorAll( 'form' );
+	const elements = document.querySelectorAll( triggers );
+	elements.forEach( element => {
+		const forms = element.querySelectorAll( 'form' );
 		forms.forEach( form => {
 			form.appendChild( modalCheckoutInput.cloneNode() );
 			form.target = iframeName;
