+++ src/modal-checkout/view.scss
@@ -1,4 +1,4 @@
-@use '../../../shared/sass/colors';
+@use '../shared/sass/colors';
 
 #newspack_modal_checkout { /* stylelint-disable-line */
 	padding: 32px;
